incorporate license information in issue templates as YAML comment
GitHub incorporates the second document in the yaml file, mis extentioned as a markdown file, as the content of the issue that is being created. When the license information was placed that it got included literally.

In this commit we write the license information as a yaml comment in side the first document